# chore: migrate to sqlite for more consistent (and faster) storage