--- a/internal/env/environment.go
+++ b/internal/env/environment.go
@@ -32,7 +32,7 @@ func ConfigFilePath() string {
 	if val := os.Getenv(EnvVarConfigPath); val != "" {
 		return val
 	}
-	return path.Join(getConfigDir(), "backrest/config.json")
+	return filepath.Join(getConfigDir(), "backrest", "config.json")
 }
 
 // DataDir
@@ -50,7 +50,7 @@ func DataDir() string {
 	}
 
 	if runtime.GOOS == "windows" {
-		return path.Join(getConfigDir(), "backrest/data")
+		return filepath.Join(getConfigDir(), "backrest", "data")
 	}
 	return path.Join(getHomeDir(), ".local/share/backrest")
 }
@@ -99,7 +99,7 @@ func getConfigDir() string {
 	if val := os.Getenv("XDG_CONFIG_HOME"); val != "" {
 		return val
 	}
-	return path.Join(getHomeDir(), ".config")
+	return filepath.Join(getHomeDir(), ".config")
 }
 
 func formatBindAddress(addr string) string {