--- a/internal/logwriter/errors.go
+++ b/internal/logwriter/errors.go
@@ -1,7 +1,0 @@
-package logwriter
-
-import "errors"
-
-var ErrFileNotFound = errors.New("file not found")
-var ErrNotFound = errors.New("entry not found")
-var ErrBadName = errors.New("bad name")