--- a/internal/logwriter/logwriter.go
+++ b/internal/logwriter/logwriter.go
@@ -1,0 +1,1 @@
+package logwriter