… go.sum
… internal/logwriter/livelog.go
… internal/logwriter/livelog_test.go
… internal/logwriter/manager.go
… internal/logwriter/manager_test.go
… internal/logwriter/rotatinglog.go
… internal/logwriter/rotatinglog_test.go
… internal/oplog/storetests/storecontract_test.go
… internal/textstore/sqlitestore_test.go
… internal/textstore/textstore.go
… internal/textstore/util.go
… pkg/restic/restic_test.go
… proto/v1/operations.proto