@garethgeorge 2024-09-20
Note: actually seeing _very_ poor performance out of the textstore implemented on sqlite. Some more thinking is needed to decide if this is the right way to go.